+++ src/test/kotlin/com/yapp/web2/domain/bookmark/service/PersonalBookmarkServiceTest.kt
@@ -1,7 +1,7 @@
 package com.yapp.web2.domain.bookmark.service
 
 import com.yapp.web2.domain.account.entity.Account
-import com.yapp.web2.domain.bookmark.entity.BookmarkDto
+import com.yapp.web2.domain.bookmark.BookmarkDto
 import com.yapp.web2.domain.bookmark.entity.PersonalBookmark
 import com.yapp.web2.domain.bookmark.repository.BookmarkInterfaceRepository
 import com.yapp.web2.domain.folder.entity.Folder
